fix: oversized messages causes clients to hang/throw exceptions (#47)
The HttpObjectAggregator might return http responses in some cases, such
as a too large body content-length being sent to the server. As the http
encoder was after this handler in the pipeline, the response couldn't be
serialized and an exception was thrown. Furthermore, there seems to be a
bug in netty which leaves the request decoder and aggregator in a
mismatched state, causing the requests sent after a first request with a
too long body to hang forever (if keep-alive is enabled). This is fixed
for now by just always closing the connection when a too large body is
detected.