fix: send bad request response on decode failure (#62)
When the client sends a bad request to the server which cannot be
decoded, for example containing an ascii control sequence, the
connection is just closed without a proper response. Now the server
returns a 400 - Bad Request response in these cases. The handling of the
failure response send has been moved into a util method to properly set
the `connection: close` and `content-length: 0` headers every time.